moved landing page to / and added a temporary image in the hero
